+++ src/components/projects/Freely.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import CaseStudyTemplate from './CaseStudyTemplate';
+import { projectsData } from '../../data/projects';
+
+function Freely() {
+  const project = projectsData.find(p => p.slug === 'freely');
+
+  return <CaseStudyTemplate project={project} />;
+}
+
+export default Freely;