+++ src/components/projects/PairupEvents.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import CaseStudyTemplate from './CaseStudyTemplate';
+import { projectsData } from '../../data/projects';
+
+function PairupEvents() {
+  const project = projectsData.find(p => p.slug === 'pairup-events');
+
+  return <CaseStudyTemplate project={project} />;
+}
+
+export default PairupEvents;