+++ src/components/projects/XingPrivateProfile.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import CaseStudyTemplate from './CaseStudyTemplate';
+import { projectsData } from '../../data/projects';
+
+function XingPrivateProfile() {
+  const project = projectsData.find(p => p.slug === 'xing-private-profile');
+
+  return <CaseStudyTemplate project={project} />;
+}
+
+export default XingPrivateProfile;